Group query endpoint and rename to spinal-case
