+++ src/main/java/com/ericsson/ei/controller/QueryController.java
@@ -8,6 +8,7 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 
@@ -23,7 +24,7 @@ public interface QueryController {
 
 
     /**
-     * The REST POST method is used to query aggregated objects with the requested criteria, which are present in the request body.
+     * This method is used to query aggregated objects with the requested criteria, which must be present in the request body. Options and filter are optional.
      * 
      */
     @RequestMapping(value = "", method = RequestMethod.POST)
@@ -31,5 +32,23 @@ public interface QueryController {
         @Valid
         @RequestBody
         QueryBody queryBody);
+
+    /**
+     * This method retrieves aggregated data on a specific aggregated object, given an ID.
+     * 
+     */
+    @RequestMapping(value = "/aggregated-object", method = RequestMethod.GET)
+    public ResponseEntity<?> getQueryAggregatedObject(
+        @RequestParam
+        String id);
+
+    /**
+     * This method returns missed notification(s) given a subscription name.
+     * 
+     */
+    @RequestMapping(value = "/missed-notifications", method = RequestMethod.GET)
+    public ResponseEntity<?> getQueryMissedNotifications(
+        @RequestParam
+        String subscriptionName);
 
 }
