+++ src/main/resources/public/raml/eiffel-intelligence.raml
@@ -36,8 +36,6 @@ mediaType: application/json
 baseUri: /
 /subscriptions: !include resourceTypes/subscription.raml
 /query: !include resourceTypes/query.raml
-/queryAggregatedObject: !include resourceTypes/query-aggregated_object.raml
-/queryMissedNotifications: !include resourceTypes/query-missed_notifications.raml
 /download: !include resourceTypes/download.raml
 /rules: !include resourceTypes/rules.raml
 /auth: !include resourceTypes/auth.raml
