+++ src/main/resources/public/raml/resourceTypes/auth.raml
@@ -14,7 +14,7 @@ get:
         body:
           application/string:
 
-/checkStatus:
+/check-status:
   get:
     description: This method checks Eiffel Intelligene backend status.
     responses:
