+++ src/main/resources/public/raml/resourceTypes/download.raml
@@ -6,7 +6,7 @@ get:
       body:
         application/string:
 
-/subscriptionsTemplate:
+/subscriptions-template:
   get:
     description: This method returns a subscription template.
     responses:
@@ -14,7 +14,7 @@ get:
         body:
           application/string:
 
-/rulesTemplate:
+/rules-template:
   get:
     description: This method returns a template for rules.
     responses:
@@ -22,7 +22,7 @@ get:
         body:
           application/string:
 
-/eventsTemplate:
+/events-template:
   get:
     description: This method returns a template for Eiffel events.
     responses:
