--- src/main/resources/public/raml/resourceTypes/query-aggregated_object.raml
@@ -1,14 +1,0 @@
-description: This class returns aggregated data given a specific ID of an aggregated object.
-get:
-  description: This method retrieves aggregated data on a specific aggregated object, given an ID.
-  queryParameters:
-    id:
-      displayName: id
-      type: string
-      description: Id of the aggregatedObject
-      example: 6acc3c87-75e0-4b6d-88f5-b1a5d4e62b44
-      required: true
-  responses:
-    200:
-      body:
-        application/json:
