--- src/main/resources/public/raml/resourceTypes/query-missed_notifications.raml
@@ -1,14 +1,0 @@
-description: This class searches for missed notifications in the database.
-get:
-  description: This method returns missed notification(s) given a subscription name.
-  queryParameters:
-    subscriptionName:
-      displayName: subscriptionName
-      type: string
-      description: SubscriptionName of the Missed Notification
-      example: Subscription_1
-      required: true
-  responses:
-    200:
-      body:
-        application/json:
