+++ src/main/resources/public/raml/resourceTypes/rules.raml
@@ -21,7 +21,7 @@ description: This class is used to test rules on Eiffel events. Testrules must b
           body:
             application/string:
 
-  /testRulePageEnabled:
+  /test-rule-page-enabled:
     get:
       description: This method checks if the possibility to test rules has been enabled in Eiffel Intelligence.
       responses:
